tools/TeleInfoReader/requirements.txt
pytest>=7.2.0
pyserial>=3.5
